+++ model/map_role_build.go
@@ -4,7 +4,7 @@ type MapRoleBuild struct {
 	Id			int			`json:"id" xorm:"id pk autoincr"`
 	RId			int			`json:"rid" xorm:"rid"`
 	Type		int8		`json:"type"`
-	Level		int			`json:"level"`
+	Level		int8		`json:"level"`
 	X			int			`json:"x"`
 	Y			int			`json:"y"`
 	Name		string		`json:"name"`
