+++ model/national_map.go
@@ -6,7 +6,7 @@ type NationalMap struct {
 	X			int		`json:"x"`
 	Y			int		`json:"y"`
 	Type		int8	`json:"type"`
-	Level		int		`json:"level"`
+	Level		int8	`json:"level"`
 	CurDurable	int		`json:"cur_durable"`
 }
